moved plotStyling to new folder
